requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.5
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-allauth==0.38.0
django-filter==2.0.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-tables2==2.0.3
django-toolbelt==0.0.1
djangorestframework==3.9.0
djangorestframework-datatables==0.4.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
Markdown==3.0.1
mysqlclient==1.3.14
oauthlib==3.0.1
pdfkit==0.6.1
Pillow==5.4.1
psycopg2==2.7.6.1
pycparser==2.19
python-dotenv==0.9.1
python-pdf==0.36
python-telegram-bot==11.1.0
python3-openid==3.1.0
pytz==2018.7
qrcode==6.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
static3==0.7.0
urllib3==1.24.1
whitenoise==4.1.2
